Use dict of slurm args instead of individual args
Add slurm kwargs

w
