+++ src/ClimaCalibrate.jl
@@ -2,6 +2,7 @@ module ClimaCalibrate
 
 include("ekp_interface.jl")
 include("model_interface.jl")
+include("slurm.jl")
 include("backends.jl")
 include("emulate_sample.jl")
