+++ src/ekp_interface.jl
@@ -86,6 +86,16 @@ path_to_ensemble_member(output_dir, iteration, member) =
     EKP.TOMLInterface.path_to_ensemble_member(output_dir, iteration, member)
 
 """
+    path_to_model_log(output_dir, iteration, member)
+
+Constructs the path to an ensemble member's forward model log for a given iteration and member number.
+"""
+path_to_model_log(output_dir, iteration, member) = joinpath(
+    path_to_ensemble_member(output_dir, iteration, member),
+    "model_log.txt",
+)
+
+"""
     path_to_iteration(output_dir, iteration)
 
 Creates the path to the directory for a specific iteration within the specified output directory.
