+++ test/Project.toml
@@ -1,6 +1,6 @@
 [deps]
-ClimaCalibrate = "4347a170-ebd6-470c-89d3-5c705c0cacc2"
 CalibrateEmulateSample = "95e48a1f-0bec-4818-9538-3db4340308e3"
+ClimaCalibrate = "4347a170-ebd6-470c-89d3-5c705c0cacc2"
 ClimaParams = "5c42b081-d73a-476f-9059-fd94b934656c"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 EnsembleKalmanProcesses = "aa8a2aa5-91d8-4396-bcef-d4f2ec43552d"
